clean up branch
Removes unused code, move code for better separation of concerns.
