+++ op-chain-ops/cmd/celo-dbmigrate/transform.go
@@ -1,11 +1,14 @@
-package celo1
+package main
 
 import (
+	"bytes"
 	"errors"
+	"fmt"
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/rlp"
 )
 
@@ -37,8 +40,8 @@ type IstanbulExtra struct {
 	ParentAggregatedSeal IstanbulAggregatedSeal
 }
 
-// RemoveIstanbulAggregatedSeal removes the aggregated seal from the header
-func RemoveIstanbulAggregatedSeal(header []byte) ([]byte, error) {
+// transformHeader removes the aggregated seal from the header
+func transformHeader(header []byte) ([]byte, error) {
 	newHeader := new(types.Header) // TODO double check on decoding type
 	err := rlp.DecodeBytes(header, newHeader)
 	if err != nil {
@@ -65,4 +68,34 @@ func RemoveIstanbulAggregatedSeal(header []byte) ([]byte, error) {
 	newHeader.Extra = append(newHeader.Extra[:IstanbulExtraVanity], payload...)
 
 	return rlp.EncodeToBytes(newHeader)
+}
+
+func hasSameHash(newHeader, oldHash []byte) (bool, common.Hash) {
+	newHash := crypto.Keccak256Hash(newHeader)
+	return bytes.Equal(oldHash, newHash.Bytes()), newHash
+}
+
+// transformBlockBody migrates the block body from the old format to the new format (works with []byte input output)
+func transformBlockBody(oldBodyData []byte) ([]byte, error) {
+	// decode body into celo-blockchain Body structure
+	// remove epochSnarkData and randomness data
+	var celoBody struct {
+		Transactions   rlp.RawValue // TODO use types.Transactions to make sure all tx are deserializable
+		Randomness     rlp.RawValue
+		EpochSnarkData rlp.RawValue
+	}
+	if err := rlp.DecodeBytes(oldBodyData, &celoBody); err != nil {
+		return nil, fmt.Errorf("failed to RLP decode body: %w", err)
+	}
+
+	// TODO create a types.BlockBody structure and encode it back to []byte
+
+	// transform into op-geth types.Body structure
+	// since Body is a slice of types.Transactions, we can just remove the randomness and epochSnarkData and add empty array for UnclesHashes
+	newBodyData, err := rlp.EncodeToBytes([]interface{}{celoBody.Transactions, nil})
+	if err != nil {
+		return nil, fmt.Errorf("failed to RLP encode body: %w", err)
+	}
+
+	return newBodyData, nil
 }
